pyproject.toml
[project]
name = "policy-advisor"
version = "0.1.0"
description = ""
authors = [
    {name = "Ridwan Yusuf",email = "yosufridwan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "chromadb (==0.4.18)",
    "deep-translator (<1.11.4)",
    "fastapi (==0.116.1)",
    "langdetect (==1.0.9)",
    "ollama (==0.5.3)",
    "pydantic (==2.11.7)",
    "pypdf2 (==3.0.1)",
    "python-dotenv (==1.1.1)",
    "redis (==6.4.0)",
    "uvicorn (==0.35.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "textract (>=1.6.5,<2.0.0)",
    "numpy (>=2.1,<3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-groq (>=0.3.8,<0.4.0)",
    "langchain-community (>=0.3.30,<0.4.0)",
    "sentence-transformers (>=5.1.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
